+++ Notes/UIKitWrapper/AttachmentTapGestureRecognizer.swift
@@ -1,0 +1,8 @@
+//
+//  AttachmentTapGestureRecognizer.swift
+//  Notes
+//
+//  Created by Lena Vadakkel on 18.01.2024.
+//
+
+import Foundation
