+++ Notes/UIKitWrapper/Extensions/UIFont.swift
@@ -1,0 +1,35 @@
+//
+//  UIFont.swift
+//  Notes
+//
+//  Created by Lena Vadakkel on 18.01.2024.
+//
+
+import Foundation
+import UIKit
+
+extension UIFont {
+	func withTraits(_ traits: UIFontDescriptor.SymbolicTraits) -> UIFont {
+		
+		// create a new font descriptor with the given traits
+		guard let fd = fontDescriptor.withSymbolicTraits(traits) else {
+			// the given traits couldn't be applied, return self
+			return self
+		}
+		
+		// return a new font with the created font descriptor
+		return UIFont(descriptor: fd, size: pointSize)
+	}
+	
+	func italics() -> UIFont {
+		return withTraits(.traitItalic)
+	}
+	
+	func bold() -> UIFont {
+		return withTraits(.traitBold)
+	}
+	
+	func boldItalics() -> UIFont {
+		return withTraits([ .traitBold, .traitItalic ])
+	}
+}
